添加CGScrollView，修改CGCycleScrollView循环滑动视图

+++ View/CustomView/View/CycleScrollView/CycleBaseScrollView/CycleSceneCategory/CGCycleScrollView+CGCondition.m
@@ -38,7 +38,7 @@
         }
     }else {
         
-        CGErrorLog(@"指定的当前索引(%li)超出索引范围(0 ~ %li)", (long)paramCurrentIndex, self.totalViewNumber);
+        CGErrorLog(@"指定的当前索引(%li)超出索引范围(0 ~ %li)", (long)paramCurrentIndex, (long)self.totalViewNumber);
     }
     return isResult;
 }
